+++ Hydrogen.Entities.Hybrid/ConvertSingleton.cs
@@ -13,7 +13,6 @@ namespace Titanium.Core
     /// <typeparam name="T1"><see cref="IComponentData"/> data type</typeparam>
     [DisallowMultipleComponent]
     [RequiresEntityConversion]
-    [RequireComponent(typeof(ConvertToEntity))]
     public abstract class ConvertSingleton<T0, T1> : MonoBehaviour, IConvertGameObjectToEntity
         where T0 : ScriptableObject
         where T1 : struct, IComponentData