+++ CHANGELOG.md
@@ -1,9 +1,15 @@
-# 0.1.1
+# Changelog
+
+## [0.1.2] - 2019-08-19
+- Simplified IConvertScriptableObjectToBlob API surface.
+- User can now access the GameObjectConversionSystem from ScriptableObjectConversionSystem's GoConversionSystem property.
+
+## [0.1.1]
 
 Minor adjustments to package naming.
 Added Initial Documentation
 
-# 0.1.0
+## [0.1.0]
 Initial commits.
 
 Config and BlobAsset/ref helpers.